requirements.txt
pandas
prophet
duckdb
streamlit
plotly
python-dotenv
requests